Tue, Feb 21, 2023, 3:14 PM +03:00

+++ interpreter/interpreter.py
@@ -1,5 +1,5 @@
 expression = (input( 'Type xyz '))
-expression = int(expression.split(' '))
+expression = float(expression.split(' '))
 
 
 print (f'{x:.2f}')
